requirements.txt
python-chess==0.14.0
h5py==2.6.0
numpy==1.11.0
six==1.10.0
